Fix/sample count

+++ src/disk/thread.rs
@@ -201,7 +201,7 @@ fn write_exec(
 ) -> std::io::Result<()> {
   if buffered.len() == 1 {
     let (p, buf) = &buffered[0];
-    metrics.disk_write_batch.record(1);
+    metrics.disk_write_batch.record(1.0);
     measure!(metrics.disk_write, backend.pwrite_all(buf, *p))?;
     return Ok(());
   }
@@ -215,7 +215,7 @@ fn write_exec(
   for chunk in buffered.chunk_by(|(a_o, a_b), (b_o, _)| a_o + a_b.len() as u64 == *b_o) {
     let (offset, mut bufs): (Vec<_>, Vec<_>) =
       chunk.iter().map(|(o, b)| (*o, *b)).unzip();
-    metrics.disk_write_batch.record(bufs.len() as u64);
+    metrics.disk_write_batch.record(bufs.len() as f64);
 
     let offset = offset[0];
     if bufs.len() == 1 {
@@ -244,7 +244,7 @@ fn flush_fdatasync(
   };
 
   let result = backend.fdatasync().map_err(|err| err.kind());
-  metrics.disk_sync_batch.record(waiting.len() as u64);
+  metrics.disk_sync_batch.record(waiting.len() as f64);
   waiting
     .drain(..)
     .for_each(|done| done.fulfill(result.map_err(Error::from)));
